requirements-test.txt
# Tests

pytest==2.9.2
pytest-tornado==0.4.5
pytest-cov==2.3.0

# Mocking and doubles
mock==2.0.0
doubles==1.2.1

# Coverage and Others
pytest-ipdb==0.1-prerelease2
pytest-timeout==1.0.0
ipdb==0.10.1

# Lint
flake8==3.0.0b2